+++ components/css.js
@@ -1,0 +1,35 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    alignItems: 'center',
+    justifyContent: 'center',
+    marginTop: 100,
+  },
+  appTitle: {
+    fontSize: 30,
+    fontWeight: 'bold',
+  },
+  input: {
+    width: 100,
+  },
+  bird: {
+    // flex: 1,
+    // alignItems: 'center',
+    // justifyContent: 'center',
+    position: 'absolute',
+    top: 50,
+    right: 30,
+    // backgroundColor: 'yellow',
+    width: 200,
+    height: 250,
+  },
+  cage: {
+    // position: 'absolute',
+    // backgroundColor: 'pink',
+    width: 400,
+    height: 550,
+  },
+});
+
+export default styles;
